Battery efficiency changes.
Changed TickTime service back to minute units. (Discovered that it fires when the minute value of the clock changes). 
Added static variables for battery and bluetooth states.
Added subscription to battery and bluetooth services so that the static state variables are only updated on a change.
Commented out call to destroy_window during deinit(). Forum comments seem to suggest it is not needed since the OS does this and in fact creates a double free error condition.